+++ lib/__init__.py
@@ -2,7 +2,7 @@
 Library package for ACM switchover automation.
 """
 
-__version__ = "1.5.10"
+__version__ = "1.5.11"
 __version_date__ = "2026-03-06"
 
 from .exceptions import (